multimodal_gemini_vs_qwen/requirements.txt
google-genai==1.12.1
openai==1.76.2
Pillow==12.0.0
protobuf==6.33.2
